Refactor GraphQL query posts field
